enchancement for physical memory management

+++ src/kernel/include/string.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "stdint.h"
+#include <stddef.h>
 
 void* memcpy(void* dst, const void* src, uint16_t num);
-void* memset(void* ptr, int value, uint16_t num);
+void* memset(void* ptr, int value, uint32_t num);
 int memcmp(const void* ptr1, const void* ptr2, uint16_t num);
